requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Restful==0.3.5
gunicorn==19.6.0
python-dotenv==0.6.0
webargs==1.3.4
marshmallow==2.7.3
PyJWT==1.4.2
passlib==1.6.5
Flask-JWT==0.3.2
PyYAML==3.12
